fix: restore UI lint budget after runs filter helpers
Move non-component helpers out of RunDataSourceFiltersPanel and
useWidgetData so react-refresh and max-lines stay within budget.
